Initial changes to drop Python2.7 support
Removes dependencies on `unicodecsv` and `six`, and limits to Python 3.3+.

Currently broken, as the stdlib `csv.reader()` and `csv.writer()` interfaces
don't support an `encoding=...` parameter which this library sets to
automatically encode input strings.

Fixing may require a custom `BytesIO` subclass to re-implement the behaivor,
or to change view classes to use `StringIO` for text mode (but that may
have other side effects).